Consistently patch all plugin.xml files in the repository
