+++ buildSrc/src/main/kotlin/toolkit-intellij-plugin.gradle.kts
@@ -1,0 +1,5 @@
+// Copyright 2024 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+import software.aws.toolkits.gradle.intellij.ToolkitIntelliJExtension
+
+project.extensions.create<ToolkitIntelliJExtension>("intellijToolkit")
